Move Global (static access) to GlobalPaginator
