+++ src/GlobalSetupInterface.php
@@ -1,0 +1,62 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace Zend\Paginator;
+
+
+interface GlobalSetupInterface
+{
+    /**
+     * Set a global config
+     *
+     * @param array|Traversable $config
+     * @throws Exception\InvalidArgumentException
+     */
+    public static function setGlobalConfig($config);
+
+    /**
+     * Returns the default scrolling style.
+     *
+     * @return  string
+     */
+    public static function getDefaultScrollingStyle();
+
+    /**
+     * Get the default item count per page
+     *
+     * @return int
+     */
+    public static function getDefaultItemCountPerPage();
+
+    /**
+     * Set the default item count per page
+     *
+     * @param int $count
+     */
+    public static function setDefaultItemCountPerPage($count);
+
+    /**
+     * Sets the default scrolling style.
+     *
+     * @param  string $scrollingStyle
+     */
+    public static function setDefaultScrollingStyle($scrollingStyle = 'Sliding');
+
+    public static function setScrollingStylePluginManager($scrollingAdapters);
+
+    /**
+     * Returns the scrolling style manager.  If it doesn't exist it's
+     * created.
+     *
+     * @return ScrollingStylePluginManager
+     */
+    public static function getScrollingStylePluginManager();
+
+
+}
