+++ src/PaginatorInterface.php
@@ -14,51 +14,6 @@ use Zend\Paginator\Adapter\AdapterInterface;
 
 interface PaginatorInterface extends Countable, IteratorAggregate
 {
-    /**
-     * Set a global config
-     *
-     * @param array|Traversable $config
-     * @throws Exception\InvalidArgumentException
-     */
-    public static function setGlobalConfig($config);
-
-    /**
-     * Returns the default scrolling style.
-     *
-     * @return  string
-     */
-    public static function getDefaultScrollingStyle();
-
-    /**
-     * Get the default item count per page
-     *
-     * @return int
-     */
-    public static function getDefaultItemCountPerPage();
-
-    /**
-     * Set the default item count per page
-     *
-     * @param int $count
-     */
-    public static function setDefaultItemCountPerPage($count);
-
-    /**
-     * Sets the default scrolling style.
-     *
-     * @param  string $scrollingStyle
-     */
-    public static function setDefaultScrollingStyle($scrollingStyle = 'Sliding');
-
-    public static function setScrollingStylePluginManager($scrollingAdapters);
-
-    /**
-     * Returns the scrolling style manager.  If it doesn't exist it's
-     * created.
-     *
-     * @return ScrollingStylePluginManager
-     */
-    public static function getScrollingStylePluginManager();
 
     /**
      * Returns the number of pages.
