+++ src/SimplePaginator.php
@@ -15,46 +15,9 @@ use Traversable;
 use Zend\Paginator\Adapter\AdapterInterface;
 use Zend\Paginator\ScrollingStyle\ScrollingStyleInterface;
 use Zend\Stdlib\ArrayUtils;
-use Zend\ServiceManager\ServiceManager;
 
 class SimplePaginator implements PaginatorInterface
 {
-
-    /**
-     * Adapter plugin manager
-     *
-     * @var AdapterPluginManager
-     */
-    protected static $adapters = null;
-
-    /**
-     * Configuration file
-     *
-     * @var array|null
-     */
-    protected static $config = null;
-
-    /**
-     * Default scrolling style
-     *
-     * @var string
-     */
-    protected static $defaultScrollingStyle = 'Sliding';
-
-    /**
-     * Default item count per page
-     *
-     * @var int
-     */
-    protected static $defaultItemCountPerPage = 10;
-
-    /**
-     * Scrolling style plugin manager
-     *
-     * @var ScrollingStylePluginManager
-     */
-    protected static $scrollingStyles = null;
-
     /**
      * Adapter
      *
@@ -88,7 +51,7 @@ class SimplePaginator implements PaginatorInterface
      *
      * @var int
      */
-    protected $itemCountPerPage = null;
+    protected $itemCountPerPage = 10;
 
     /**
      * Number of pages
@@ -113,111 +76,6 @@ class SimplePaginator implements PaginatorInterface
     protected $pages = null;
 
     /**
-     * Set a global config
-     *
-     * @param array|Traversable $config
-     * @throws Exception\InvalidArgumentException
-     */
-    public static function setGlobalConfig($config)
-    {
-        if ($config instanceof Traversable) {
-            $config = ArrayUtils::iteratorToArray($config);
-        }
-        if (!is_array($config)) {
-            throw new Exception\InvalidArgumentException(__METHOD__ . ' expects an array or Traversable');
-        }
-
-        static::$config = $config;
-
-        if (isset($config['scrolling_style_plugins'])
-            && null !== ($adapters = $config['scrolling_style_plugins'])
-        ) {
-            static::setScrollingStylePluginManager($adapters);
-        }
-
-        $scrollingStyle = isset($config['scrolling_style']) ? $config['scrolling_style'] : null;
-
-        if ($scrollingStyle !== null) {
-            static::setDefaultScrollingStyle($scrollingStyle);
-        }
-    }
-
-    /**
-     * Returns the default scrolling style.
-     *
-     * @return  string
-     */
-    public static function getDefaultScrollingStyle()
-    {
-        return static::$defaultScrollingStyle;
-    }
-
-    /**
-     * Get the default item count per page
-     *
-     * @return int
-     */
-    public static function getDefaultItemCountPerPage()
-    {
-        return static::$defaultItemCountPerPage;
-    }
-
-    /**
-     * Set the default item count per page
-     *
-     * @param int $count
-     */
-    public static function setDefaultItemCountPerPage($count)
-    {
-        static::$defaultItemCountPerPage = (int) $count;
-    }
-
-    /**
-     * Sets the default scrolling style.
-     *
-     * @param  string $scrollingStyle
-     */
-    public static function setDefaultScrollingStyle($scrollingStyle = 'Sliding')
-    {
-        static::$defaultScrollingStyle = $scrollingStyle;
-    }
-
-    public static function setScrollingStylePluginManager($scrollingAdapters)
-    {
-        if (is_string($scrollingAdapters)) {
-            if (!class_exists($scrollingAdapters)) {
-                throw new Exception\InvalidArgumentException(sprintf(
-                    'Unable to locate scrolling style plugin manager with class "%s"; class not found',
-                    $scrollingAdapters
-                ));
-            }
-            $scrollingAdapters = new $scrollingAdapters(new ServiceManager);
-        }
-        if (!$scrollingAdapters instanceof ScrollingStylePluginManager) {
-            throw new Exception\InvalidArgumentException(sprintf(
-                'Pagination scrolling-style manager must extend ScrollingStylePluginManager; received "%s"',
-                (is_object($scrollingAdapters) ? get_class($scrollingAdapters) : gettype($scrollingAdapters))
-            ));
-        }
-        static::$scrollingStyles = $scrollingAdapters;
-    }
-
-    /**
-     * Returns the scrolling style manager.  If it doesn't exist it's
-     * created.
-     *
-     * @return ScrollingStylePluginManager
-     */
-    public static function getScrollingStylePluginManager()
-    {
-        if (static::$scrollingStyles === null) {
-            static::$scrollingStyles = new ScrollingStylePluginManager(new ServiceManager);
-        }
-
-        return static::$scrollingStyles;
-    }
-
-    /**
      * Constructor.
      *
      * @param AdapterInterface|AdapterAggregateInterface $adapter
@@ -234,22 +92,6 @@ class SimplePaginator implements PaginatorInterface
                 'Zend\Paginator only accepts instances of the type ' .
                 'Zend\Paginator\Adapter\AdapterInterface or Zend\Paginator\AdapterAggregateInterface.'
             );
-        }
-
-        $config = static::$config;
-
-        if (!empty($config)) {
-            $setupMethods = ['ItemCountPerPage', 'PageRange'];
-
-            foreach ($setupMethods as $setupMethod) {
-                $key   = strtolower($setupMethod);
-                $value = isset($config[$key]) ? $config[$key] : null;
-
-                if ($value !== null) {
-                    $setupMethod = 'set' . $setupMethod;
-                    $this->$setupMethod($value);
-                }
-            }
         }
     }
 
@@ -406,10 +248,6 @@ class SimplePaginator implements PaginatorInterface
      */
     public function getItemCountPerPage()
     {
-        if (empty($this->itemCountPerPage)) {
-            $this->itemCountPerPage = static::getDefaultItemCountPerPage();
-        }
-
         return $this->itemCountPerPage;
     }
 
@@ -513,11 +351,12 @@ class SimplePaginator implements PaginatorInterface
     /**
      * Returns the page collection.
      *
-     * @param  string $scrollingStyle Scrolling style
+     * @param  ScrollingStyleInterface $scrollingStyle Scrolling style
      * @return \stdClass
      */
     public function getPages($scrollingStyle = null)
     {
+
         if ($this->pages === null) {
             $this->pages = $this->_createPages($scrollingStyle);
         }
@@ -657,10 +496,6 @@ class SimplePaginator implements PaginatorInterface
      */
     protected function _loadScrollingStyle($scrollingStyle = null)
     {
-        if ($scrollingStyle === null) {
-            $scrollingStyle = static::$defaultScrollingStyle;
-        }
-
         switch (strtolower(gettype($scrollingStyle))) {
             case 'object':
                 if (!$scrollingStyle instanceof ScrollingStyleInterface) {
@@ -670,9 +505,6 @@ class SimplePaginator implements PaginatorInterface
                 }
 
                 return $scrollingStyle;
-
-            case 'string':
-                return static::getScrollingStylePluginManager()->get($scrollingStyle);
 
             case 'null':
                 // Fall through to default case
